Line compare register writes are obviously wrong:
Register 0x33 is written with value 0x05 using mask 0x06
It causes left side of the screen to be garbled and shifted right
at least on K8M890CE.

Register 0x09 is written using mask 0xff, overwriting bit 5
(which is bit 9 of vblank_start register).
It causes blue horizontal bar on top of screen and blank bar in the middle
at least on K8M890CE.

Fix the writes to match viafb.

Signed-off-by: Ondrej Zary <linux@rainbow-software.org>