Fix for FP only use (laptop) not working correctly
It was observed that if analog VGA and FP (Flat Panel) were
connected during boot time, FP worked, but if only FP was connected,
FP was not displaying the correct image. The code currently has issues
when IGA1 is assigned to FP. Previously IGA1 or IGA2 can be assigned
to FP. Now, only IGA2 will be assigned to FP. This bug fix solves
OpenChrome DRM for VIA Technologies Chrome IGP laptop usability issue.

Signed-off-by: Kevin Brace <kevinbrace@gmx.com>