Version bumped to 3.0.44
Made lots of changes to the code in order to be able to compile and
work with Linux 4.13 rc5 kernel. OpenChrome DRM no longer crashes when
booting Xubuntu 16.04.3 with Linux 4.13 rc5 kernel on HP 2133 mini-note.
Due to many issues encountered during the porting work, it took roughly
two months of intense effort to get here (July to September 2017).
Changed the driver author to OpenChrome Project (i.e., dropped ‘The’
portion).

Signed-off-by: Kevin Brace <kevinbrace@gmx.com>