Version bumped to 3.0.45
Made improvements in how FP is detected and initialized. DRM / X Server
will still not be able to reinitialize the screen properly after standby
resume, but the improvements made by this version can restore FP only
once when the user switches to a VT (Virtual Terminal) from a blank
screen. For the developer, having a working VT is far better than a blank
screen for debugging purposes.

Signed-off-by: Kevin Brace <kevinbrace@gmx.com>