Discontinue the use of DMI for detecting FP availability
The use of DMI (Desktop Management Interface) is awkward, so it will
be discontinued. The detection of Quanta IL1 netbook is done via DMI,
but this will be changed to PCI Subvendor and Subdevice ID in the near
future. Quanta IL1 netbook’s FP is connected to DVP1 since it is a TTL
interface FP, not LVDS interface FP. Because of this, a special flag
register will be used to properly recognize / control the FP of this
computer.

Signed-off-by: Kevin Brace <kevinbrace@gmx.com>