Initial code for supporting standby resume
After struggling for more than a year trying to figure out how to fix
standby resume (ACPI S3 State resume), finally figured it out what was
wrong and how to implement it correctly. Obviously, more fixes have to
be put in for standby resume to be reliable, but at least for now, it
is working on Xubuntu 16.04.3 LTS with HP 2133 mini-note (VN896 chipset)
as long as lid close event is avoided.

Signed-off-by: Kevin Brace <kevinbrace@gmx.com>