drm/openchrome: Convert to atomic mode setting
The conversion to atomic mode setting was done in a way to preserve
the old "legacy" mode setting code as much as possible.  The changes
between the two are fairly minimal considering the difference in
semantics.  The suspend and resume code was borrowed from AST
(ASpeed Technology) DRM, and applied to the existing power management
code.  This code update was necessary for suspend and resume to work
properly with atomic mode setting.

Signed-off-by: Kevin Brace <kevinbrace@gmx.com>