drm/via: Fix for DVI on AGP based UniChrome (Pro) graphics
Even with the use of strapping resistors, it is quite challenging to
automatically detect an external flat panel, TV encoder, or DVI
transmitter properly.  It turns out that there are complicated rules
to figure out if DVP0 is configured for DVI transmitter use.  The
configuration is not well documented even inside VIA Technologies
official documentation.  The code was tested on Wyse V10LE thin
client.  This model does not support ACPI S3 State (Suspend to RAM),
so it is not clear if the screen will recover properly when coming out
of standby.

Reported-by: Eric Kudzin <knoppix1337@yahoo.com>
Signed-off-by: Kevin Brace <kevinbrace@bracecomputerlab.com>